example/case2_memo_agent/pyproject.toml
[project]
name = "case2-memo-agent"
version = "0.1.0"
description = "FastMCP server that exposes memo FastAPI endpoints as MCP tools"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.4.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.8",
    "langgraph>=0.2.51",
    "langchain-openai>=0.2.5",
    "langchain-mcp>=0.1.1",
    "langchain-mcp-adapters>=0.1.13",
]

[tool.uv]
package = true
